✨ 계정 초기 설정 페이지 구현
병합할 때 오타가 있었네요...
